Adding BabelJS on init process
